LayoutTests/http/tests/security/contentSecurityPolicy/media-src-track-block-expected.txt

CONSOLE MESSAGE: Refused to load the media 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/track.vtt' because it violates the following Content Security Policy directive: "media-src 'none'".

ALERT: PASS
